requirements.txt
opencv-python==4.8.1.78
numpy==1.26.1
requests==2.31.0
Pillow==10.1.0
pygame==2.5.2
pasimple==0.0.2